Delete states

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -16,7 +16,7 @@ if __name__ == "__main__":
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    states_del = [state for state in session.query(State) if "a" in state.name]
-    for state in states_del:
-        session.delete(state)
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
